glg-webapp/requirements.txt
matplotlib==3.5.1
streamlit==1.9.0
pandas==1.2.4
numpy==1.21.5
boto3
st-annotated-text
